refactor: add lazy create-object/comment-config helpers and begin plugin extraction
- Add `addCreateObjectModule` and `addCommentConfigModule` helpers to `AppPlugin`
  namespace in `app-toolkit/src/plugin.ts`
- Extract BoardPlugin inline create-object module to lazy `capabilities/create-object.ts`
- Create template plugin lazy create-object capability (partial - agent in progress)
