+++ packages/plugins/plugin-board/src/BoardPlugin.tsx
@@ -2,40 +2,17 @@
 // Copyright 2023 DXOS.org
 //
 
-import * as Effect from 'effect/Effect';
-
-import { Capability, Plugin } from '@dxos/app-framework';
-import { AppActivationEvents, AppPlugin } from '@dxos/app-toolkit';
-import { Operation } from '@dxos/compute';
-import { SpaceOperation } from '@dxos/plugin-space/operations';
-import { SpaceCapabilities, type CreateObject } from '@dxos/plugin-space/types';
+import { Plugin } from '@dxos/app-framework';
+import { AppPlugin } from '@dxos/app-toolkit';
 import { translations as boardTranslations } from '@dxos/react-ui-board/translations';
 
-import { ReactSurface } from '#capabilities';
+import { CreateObject, ReactSurface } from '#capabilities';
 import { meta } from '#meta';
 import { translations } from '#translations';
 import { Board } from '#types';
 
 export const BoardPlugin = Plugin.define(meta).pipe(
-  Plugin.addModule({
-    id: 'create-object',
-    activatesOn: AppActivationEvents.SetupMetadata,
-    activate: Effect.fnUntraced(function* () {
-      return Capability.contributes(SpaceCapabilities.CreateObjectEntry, {
-        id: Board.Board.typename,
-        createObject: ((props, options) =>
-          Effect.gen(function* () {
-            const object = Board.makeBoard(props);
-            return yield* Operation.invoke(SpaceOperation.AddObject, {
-              object,
-              target: options.target,
-              hidden: true,
-              targetNodeId: options.targetNodeId,
-            });
-          })) satisfies CreateObject,
-      });
-    }),
-  }),
+  AppPlugin.addCreateObjectModule({ activate: CreateObject }),
   AppPlugin.addSchemaModule({ schema: [Board.Board] }),
   AppPlugin.addSurfaceModule({ activate: ReactSurface }),
   AppPlugin.addTranslationsModule({ translations: [...translations, ...boardTranslations] }),
