+++ packages/plugins/plugin-board/src/capabilities/create-object.ts
@@ -1,0 +1,30 @@
+//
+// Copyright 2023 DXOS.org
+//
+
+import * as Effect from 'effect/Effect';
+
+import { Capability } from '@dxos/app-framework';
+import { Operation } from '@dxos/compute';
+import { SpaceOperation } from '@dxos/plugin-space/operations';
+import { SpaceCapabilities, type CreateObject } from '@dxos/plugin-space/types';
+
+import { Board } from '#types';
+
+export default Capability.makeModule(
+  Effect.fnUntraced(function* () {
+    return Capability.contributes(SpaceCapabilities.CreateObjectEntry, {
+      id: Board.Board.typename,
+      createObject: ((props, options) =>
+        Effect.gen(function* () {
+          const object = Board.makeBoard(props);
+          return yield* Operation.invoke(SpaceOperation.AddObject, {
+            object,
+            target: options.target,
+            hidden: true,
+            targetNodeId: options.targetNodeId,
+          });
+        })) satisfies CreateObject,
+    });
+  }),
+);
