+++ packages/plugins/plugin-board/src/capabilities/index.ts
@@ -4,4 +4,5 @@
 
 import { Capability } from '@dxos/app-framework';
 
+export const CreateObject = Capability.lazy('CreateObject', () => import('./create-object'));
 export const ReactSurface = Capability.lazy('ReactSurface', () => import('./react-surface'));
