+++ packages/plugins/plugin-template/src/capabilities/create-object.ts
@@ -1,0 +1,30 @@
+//
+// Copyright 2023 DXOS.org
+//
+
+import * as Effect from 'effect/Effect';
+
+import { Capability } from '@dxos/app-framework';
+import { Operation } from '@dxos/compute';
+import { SpaceOperation } from '@dxos/plugin-space/operations';
+import { SpaceCapabilities, type CreateObject } from '@dxos/plugin-space/types';
+
+import { Template } from '#types';
+
+export default Capability.makeModule(
+  Effect.fnUntraced(function* () {
+    return Capability.contributes(SpaceCapabilities.CreateObjectEntry, {
+      id: Template.Data.typename,
+      createObject: ((props, options) =>
+        Effect.gen(function* () {
+          const object = Template.make(props);
+          return yield* Operation.invoke(SpaceOperation.AddObject, {
+            object,
+            target: options.target,
+            hidden: true,
+            targetNodeId: options.targetNodeId,
+          });
+        })) satisfies CreateObject,
+    });
+  }),
+);
